feat(dashboard): make overview the default project landing
The bare project route redirects to /overview and the Apps list moves
to its own /apps route inside the project group. Project cards,
crumbs, and post-create/post-subscribe returns repoint accordingly;
crumb rendering gains exhaustiveness guards.
